A bit more cleanup/refactoring, clear out some warnings
