+++ lib/chat_api/messages/helpers.ex
@@ -1,0 +1,20 @@
+defmodule ChatApi.Messages.Helpers do
+  @moduledoc """
+  Helpers for Messages context
+  """
+
+  alias ChatApi.Messages.Message
+
+  @spec get_conversation_topic(Message.t()) :: binary()
+  def get_conversation_topic(%{conversation_id: conversation_id} = _message),
+    do: "conversation:" <> conversation_id
+
+  @spec format(Message.t()) :: map()
+  def format(%Message{} = message),
+    do: ChatApiWeb.MessageView.render("expanded.json", message: message)
+
+  @spec get_message_type(Message.t()) :: atom()
+  def get_message_type(%Message{customer_id: nil}), do: :agent
+  def get_message_type(%Message{user_id: nil}), do: :customer
+  def get_message_type(_message), do: :unknown
+end
