+++ lib/chat_api/messages/notification.ex
@@ -1,0 +1,99 @@
+defmodule ChatApi.Messages.Notification do
+  @moduledoc """
+  Notification handlers for messages
+  """
+
+  alias ChatApi.EventSubscriptions
+  alias ChatApi.Messages.{Helpers, Message}
+
+  require Logger
+
+  @spec send_webhook_notifications(binary(), map()) :: [Tesla.Env.result()]
+  def send_webhook_notifications(account_id, payload) do
+    EventSubscriptions.notify_event_subscriptions(account_id, %{
+      "event" => "message:created",
+      "payload" => payload
+    })
+  end
+
+  @spec broadcast_to_conversation!(Message.t()) :: Message.t()
+  def broadcast_to_conversation!(%Message{} = message) do
+    message
+    |> Helpers.get_conversation_topic()
+    |> ChatApiWeb.Endpoint.broadcast!("shout", Helpers.format(message))
+
+    message
+  end
+
+  @spec notify(Message.t(), atom()) :: Message.t()
+  def notify(
+        %Message{body: _body, conversation_id: _conversation_id} = message,
+        :slack
+      ) do
+    Logger.info("Sending notification: :slack")
+    # TODO: should we just pass in the message struct here?
+    Task.start(fn ->
+      ChatApi.Slack.Notifications.notify_primary_channel(message)
+    end)
+
+    message
+  end
+
+  def notify(%Message{account_id: account_id} = message, :webhooks) do
+    Logger.info("Sending notification: :webhooks")
+    # TODO: use Oban instead?
+    Task.start(fn ->
+      send_webhook_notifications(account_id, Helpers.format(message))
+    end)
+
+    message
+  end
+
+  def notify(%Message{} = message, :new_message_email) do
+    Logger.info("Sending notification: :new_message_email")
+    # TODO: use Oban instead?
+    Task.start(fn ->
+      # TODO: update params to just accept a full `message` object/struct,
+      # so that we can include some info about the customer in the email as well
+      ChatApi.Emails.send_new_message_alerts(message)
+    end)
+
+    message
+  end
+
+  def notify(%Message{} = message, :conversation_reply_email) do
+    Logger.info("Sending notification: :conversation_reply_email")
+    # 2 minutes (TODO: make this configurable?)
+    schedule_in = 2 * 60
+    formatted = Helpers.format(message)
+
+    # TODO: not sure the best way to handle this, but basically we want to only
+    # enqueue the latest message to trigger an email if it remains unseen for 2 mins
+    ChatApi.Workers.SendConversationReplyEmail.cancel_pending_jobs(formatted)
+
+    %{message: formatted}
+    |> ChatApi.Workers.SendConversationReplyEmail.new(schedule_in: schedule_in)
+    |> Oban.insert()
+
+    message
+  end
+
+  # TODO: come up with a better name... it's not super clear what `slack_support_threads` means!
+  def notify(%Message{} = message, :slack_support_threads) do
+    Logger.info("Sending notification: :slack_support_threads")
+    # TODO: should we just pass in the message struct here?
+    Task.start(fn ->
+      ChatApi.Slack.Notifications.notify_auxiliary_threads(message)
+    end)
+
+    message
+  end
+
+  def notify(message, type) do
+    Logger.error(
+      "Unrecognized notification type #{inspect(type)} for message #{inspect(message)}"
+    )
+
+    message
+  end
+end
