+++ test/chat_api/slack_test.exs
@@ -21,15 +21,15 @@ defmodule ChatApi.SlackTest do
       {:ok, conversation: conversation, account: account, customer: customer, thread: thread}
     end
 
-    test "is_valid_access_token?/1 checks the validity of an access token" do
-      assert Slack.is_valid_access_token?("invalid") == false
-      assert Slack.is_valid_access_token?("xoxb-xxx-xxxxx-xxx") == true
+    test "Token.is_valid_access_token?/1 checks the validity of an access token" do
+      assert Slack.Token.is_valid_access_token?("invalid") == false
+      assert Slack.Token.is_valid_access_token?("xoxb-xxx-xxxxx-xxx") == true
     end
 
-    test "get_message_text/1 returns subject for initial slack thread",
+    test "Helpers.get_message_text/1 returns subject for initial slack thread",
          %{conversation: conversation, customer: customer} do
       text =
-        Slack.get_message_text(%{
+        Slack.Helpers.get_message_text(%{
           customer: customer,
           text: "Test message",
           conversation_id: conversation.id,
@@ -42,9 +42,9 @@ defmodule ChatApi.SlackTest do
       assert String.contains?(text, "Reply to this thread to start chatting")
     end
 
-    test "get_message_text/1 returns subject for slack reply",
+    test "Helpers.get_message_text/1 returns subject for slack reply",
          %{conversation: conversation, customer: customer, thread: thread} do
-      assert Slack.get_message_text(%{
+      assert Slack.Helpers.get_message_text(%{
                text: "Test message",
                conversation_id: conversation.id,
                customer: customer,
@@ -53,7 +53,7 @@ defmodule ChatApi.SlackTest do
              }) ==
                "*:female-technologist: Agent*: Test message"
 
-      assert Slack.get_message_text(%{
+      assert Slack.Helpers.get_message_text(%{
                text: "Test message",
                conversation_id: conversation.id,
                customer: customer,
@@ -63,7 +63,7 @@ defmodule ChatApi.SlackTest do
                "*:wave: #{customer.email}*: Test message"
 
       assert_raise ArgumentError, fn ->
-        Slack.get_message_text(%{
+        Slack.Helpers.get_message_text(%{
           text: "Test message",
           conversation_id: conversation.id,
           customer: customer,
@@ -73,7 +73,7 @@ defmodule ChatApi.SlackTest do
       end
     end
 
-    test "get_message_payload/2 returns payload for initial slack thread",
+    test "Helpers.get_message_payload/2 returns payload for initial slack thread",
          %{customer: customer, thread: thread} do
       text = "Hello world"
       customer_email = "*Email:*\n#{customer.email}"
@@ -111,14 +111,14 @@ defmodule ChatApi.SlackTest do
                ],
                "channel" => ^channel
              } =
-               Slack.get_message_payload(text, %{
+               Slack.Helpers.get_message_payload(text, %{
                  channel: channel,
                  customer: customer,
                  thread: nil
                })
     end
 
-    test "get_message_payload/2 returns payload for slack reply",
+    test "Helpers.get_message_payload/2 returns payload for slack reply",
          %{thread: thread} do
       text = "Hello world"
       ts = thread.slack_thread_ts
@@ -129,33 +129,33 @@ defmodule ChatApi.SlackTest do
                "text" => ^text,
                "thread_ts" => ^ts
              } =
-               Slack.get_message_payload(text, %{
+               Slack.Helpers.get_message_payload(text, %{
                  channel: channel,
                  thread: thread,
                  customer: nil
                })
     end
 
-    test "extract_slack_conversation_thread_info/1 extracts thread info from slack response" do
+    test "Helpers.extract_slack_conversation_thread_info/1 extracts thread info from slack response" do
       channel = "bots"
       ts = "1234.56789"
       response = %{body: %{"ok" => true, "channel" => channel, "ts" => ts}}
 
       assert %{slack_channel: ^channel, slack_thread_ts: ^ts} =
-               Slack.extract_slack_conversation_thread_info(response)
+               Slack.Helpers.extract_slack_conversation_thread_info(response)
     end
 
-    test "extract_slack_conversation_thread_info/1 raises if the slack response has ok=false" do
+    test "Helpers.extract_slack_conversation_thread_info/1 raises if the slack response has ok=false" do
       response = %{body: %{"ok" => false}}
 
       assert capture_log(fn ->
                assert_raise RuntimeError, fn ->
-                 Slack.extract_slack_conversation_thread_info(response)
+                 Slack.Helpers.extract_slack_conversation_thread_info(response)
                end
              end) =~ "Error sending Slack message"
     end
 
-    test "create_new_slack_conversation_thread/2 creates a new thread and assigns the primary user",
+    test "Helpers.create_new_slack_conversation_thread/2 creates a new thread and assigns the primary user",
          %{conversation: conversation, account: account} do
       %{account_id: account_id, id: id} = conversation
       primary_user = insert(:user, account: account)
@@ -163,7 +163,7 @@ defmodule ChatApi.SlackTest do
       ts = "1234.56789"
       response = %{body: %{"ok" => true, "channel" => channel, "ts" => ts}}
 
-      {:ok, thread} = Slack.create_new_slack_conversation_thread(id, response)
+      {:ok, thread} = Slack.Helpers.create_new_slack_conversation_thread(id, response)
 
       assert %SlackConversationThreads.SlackConversationThread{
                slack_channel: ^channel,
@@ -177,7 +177,7 @@ defmodule ChatApi.SlackTest do
       assert conversation.assignee_id == primary_user.id
     end
 
-    test "fetch_valid_user/1 reject disabled users and fetch the oldest user.",
+    test "Helpers.fetch_valid_user/1 reject disabled users and fetch the oldest user.",
          %{account: account} do
       {:ok, disabled_user} =
         insert(:user, account: account)
@@ -190,17 +190,17 @@ defmodule ChatApi.SlackTest do
       secondary_user = insert(:user, account: account)
 
       users = [disabled_user, secondary_user, primary_user]
-      assert primary_user.id === Slack.fetch_valid_user(users)
+      assert primary_user.id === Slack.Helpers.fetch_valid_user(users)
     end
 
-    test "create_new_slack_conversation_thread/2 raises if no primary user exists",
+    test "Helpers.create_new_slack_conversation_thread/2 raises if no primary user exists",
          %{conversation: conversation} do
       channel = "bots"
       ts = "1234.56789"
       response = %{body: %{"ok" => true, "channel" => channel, "ts" => ts}}
 
       assert_raise RuntimeError, fn ->
-        Slack.create_new_slack_conversation_thread(conversation.id, response)
+        Slack.Helpers.create_new_slack_conversation_thread(conversation.id, response)
       end
     end
   end
